refactor posts to own module
